+++ src/client/java/com/orelizards/client/OreLizardModel.java
@@ -7,13 +7,13 @@ import software.bernie.geckolib.model.GeoModel;
 
 public class OreLizardModel extends GeoModel<OreLizardEntity> {
 	private static final ResourceLocation MODEL =
-			new ResourceLocation(OreLizardsMod.MOD_ID, "geo/entity/ore_lizard.geo.json");
+			ResourceLocation.fromNamespaceAndPath(OreLizardsMod.MOD_ID, "geo/entity/ore_lizard.geo.json");
 	private static final ResourceLocation TEXTURE =
-			new ResourceLocation(OreLizardsMod.MOD_ID, "textures/entity/ore_lizard.png");
+			ResourceLocation.fromNamespaceAndPath(OreLizardsMod.MOD_ID, "textures/entity/ore_lizard.png");
 	private static final ResourceLocation TEXTURE_DEEPSLATE =
-			new ResourceLocation(OreLizardsMod.MOD_ID, "textures/entity/ore_lizard_deepslate.png");
+			ResourceLocation.fromNamespaceAndPath(OreLizardsMod.MOD_ID, "textures/entity/ore_lizard_deepslate.png");
 	private static final ResourceLocation ANIMATIONS =
-			new ResourceLocation(OreLizardsMod.MOD_ID, "animations/entity/ore_lizard.animation.json");
+			ResourceLocation.fromNamespaceAndPath(OreLizardsMod.MOD_ID, "animations/entity/ore_lizard.animation.json");
 
 	@Override
 	public ResourceLocation getModelResource(OreLizardEntity animatable) {